requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
Flask-RESTful
Flask-JWT-Extended
mysql-connector
boto3
email-validator
passlib
requests